REST: Explicitly define fields
While a little more verbose, this ensures the fields that are exposed
by the API remain consistent. This also resolves an issues with
django-rest-framework 3.5, which expects either 'fields' or 'excludes'
to be defined on all serializers.

The Meta class for the Project and Patch serializers are moved to the
bottom of the serializers for consistency's sake.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Cc: Andy Doan <andy.doan@linaro.org>