REST: enable token authentication
Token authentication is generally viewed as a more secure option for API
authentication than storing a username and password.

Django REST Framework gives us a TokenAuthentication class and an authtoken
app that we can use to generate random tokens and authenticate to API
endpoints.

Enable DRF's token support and add options to the user profile view to view
or regenerate tokens.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>

---

This is an RFC; I haven't written any tests or documentation, UI's a bit
ugly, need to split patches.