docker: use libeatmydata to speed up MySQL
Creating or tearing down the patchwork database can be somewhat
time-consuming, which gets annoying when you're doing development. In
particular, when running the test suite, we recreate the database a lot. It
would be nice if this could be a bit faster for developers with slow
storage.

libeatmydata[0] is a wonderful library that makes your disk I/O super fast
by getting rid of silly things like fsync(), because writing data safely is
overrated!

Install libeatmydata in the MySQL docker container so that mysqld can use
it.

On my system, this cuts down the time to run 'tox XXXXXXXXXXXXX' from
around XXXX to XXXX.

[0] https://github.com/stewartsmith/libeatmydata

Cc: Daniel Axtens <dja@axtens.net>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>

---

TODO: someone might actually want to use docker-compose to run a prod
instance of patchwork, so make this optional...