[PW3] Drop Python 2.7
It will be unsupported by the time of release anyway.

Signed-off-by: Daniel Axtens <dja@axtens.net>

---

v2, 3: remove embarassing rebasing errors

v4: rebase on top of Python 3.8 support patch

My plan for this at the moment is to keep the pw3 topic branch looking
like:

 master ... [fixes] ... drop py2 ... [features] ... xmlrpc

so as we do development, fixes will sit before dropping py2, and
features will sit after droppng py2. This will enable fixes to be
easily backported, but means that features don't have to develop
against both py2 and py3. I'm putting xmlrpc right at the end, so
we can take it or not depending on how we go with the tools.

I want to be able to take things like Andrew's forthcoming Django3
series - which is incompatible with Python 2 - and this seems like the
simplest way to meet everyone's different goals.

Once we get 2.2 out the door and we go back to just developing on
master, we can think about how we want to handle this going forward.