[PW3] Add Django 3.0 support
Now that we've dropped Python 2, we can get ready for Django 3.0, which is
still in beta.

Add Django 3.0b1 as a tox environment. Add a release note for Django 3.0
support, as it will be released by the time Patchwork 3.0 is out.

Closes: #311 ("Django 3.0 support")
Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>

---

v1->v2:
- rebase on top of python 3.8 patch
- add release note as suggested by dja

Not adding upper bounds on the version numbers until 3.0 actually drops.

I'm considering this as closing the issue for Django 3.0 support, as now
that Django 3.0 is in beta it's unlikely to get further breaking changes
and we'll catch anything as we test it.